Adjust Python interpreter in shebangs.
The repository used both: /usr/bin/python and /usr/bin/env python.

The commit adjust the interpreter selection to /usr/bin/env python, such
that e.g. virtuelenvs are respected correctly.

This is needed when the system's Python is at version 3 and one wants a
local virtualenv for Python 2 or the tempita package.

Signed-off-by: Daniel J. Hofmann <daniel@trvx.org>