Clean up data-flow in resolveVAddr.
Initialise the 'ret' struct from a constant zero struct to ensure that no
uninitialised data is returned. Returning uninitialized data creates headaches
for binary verification.

This function is inlined in builds I have looked at, and the 'ret' struct is
never built, so this change has zero impact on the binary.