ia32: Rework VCPU interface to have a syscall that explicitly enters vt-x mode
The purpose of this change is to not pretend that a VCPU is equivalent to a regular
seL4 thread, as this leads to various state management inefficiencies. A thread
is now considered to have an additional mode (running a VM) that it can explicitly
switch to, this mode also has its own state. This allows for efficient state
management and for the user VM to control precisely when the VMM is running