Add a config option for the word size of the current platform.
This option is not necessary at compile-time as it is subsumed by the
compiler's #defines, but it is useful for code generators such as CAmkES to
know, at generation-time, the native word size of the platform they are
generating code for. Existing mechanisms either hard code a 32-bit platform or
use a set of fragile heuristics to work out the word size. Note that this
option is set automatically and not visible from menuconfig.