Add gsCNode assertion to locateSlot; clarify.
Splits the locateSlot operation into locateSlotCNode (checks that
there is indeed a CNode present) and locateSlotTCB (does not). Also
add an alias for use where the CNode pointer is immediately grabbed
from a CNode cap.

The IRQ node is treated specially in the proofs at every level, so
another variant avoids any checking.

A new assertion at retype time is to ensure that retyping either
removes CNode arrays or preserves them, but doesn't cut into them.
This is done with an assertion because it's trivial to prove in the
abstract spec (where CNode arrays are single objects) and painful
elsewhere.