x86: Move invalidateTLBEntry in unmapPage
invalidateTLBEntry is currently called twice in unmapPage
Once behind a guard that only performs the invalidate if
the active vspace is the same as the one that contains the
mapping, and then a second time at the end of unmapPage
without any guard at all.

This commit removes the unguarded version and replaces it
with the guarded one from earlier in the function.