Kconfig: Adjust default root task CNode size to align with manual.
The default CNode size of the root task as described by the manual and the
in-kernel headers is 12 bits. This is consistent when building the kernel
standalone, but in the context of an seL4 project the Kconfig defaults are
applied. This means that the default CNode size of the root task was being set
to 16 bits when building a full kernel and userspace. This commit adjusts the
Kconfig default to be the same as the other defaults to reduce future
confusion.