SELFOUR-480: replace SaveCaller with SwapCaller
This commit changes seL4_CNode_SaveCaller to
seL4_CNode_SwapCaller, which is a much more
powerful mechanism. Instead of saving the reply
cap to a cslot and putting a null cap in the
reply slot, this function can swap an earlier saved
reply cap into the tcb's slot, whilst also saving the current
one. Either candidate for swapping can be null, but cannot be
any other cap.

The use case for this change is a server (e.g a user-level scheduler)
that does not reply
to clients immediately: it picks a client to resume for each
request. This call allows the server to invoke
seL4_CNode_SwapCaller to save the reply cap for the previous client,
and store the reply cap for the client to be resumed in the
tcb reply cap slot. Then the server can call seL4_ReplyRecv.
This improves performance by allowing the server to hit the fastpath,
and allowing the server to call the same kernel function instead
of using seL4_SignalRecv, which cannot hit the fastpath and adds more
conditionals / more instructions executed.