tools: emit error if first invocation conditional
Placing a config condition on the first define
in sel4arch.xml will fail as currently the first define
must start at the last define + 1 of the previous
invocation enum.

If this is ever required, autoconf.h will need to be passed
to the script so it can check config variables itself.