x86: Use setCurrentPD instead of write_cr3
setCurrentPD caches a copy of the value that is written to cr3. This
value is used by other operations (such as invalidateTLB) to avoid
having to read cr3 in order to rewrite it.

This fixes a bug where if benchmarking is enabled invalidateTLB will be
called in boot code before setCurrentPD has ever been called, which will
result in writing a garbage value to cr3