Makefile: Prevent macro redefinition warning
These defines were added to the Makefile so that standalone debug builds
would have printing/debug support. However, in the case of a non standalone
kernel build these are redefinitions, which is fine except for
CONFIG_USER_STACK_TRACE_LENGTH, which is possibly getting defined as
something different reuslting in a compilation warning.

This commit moves these definitions to open occur if a standalone
build is actually happening.