x64: Use explicit interrupt stack
This commit uses the Interrupt Stack Table introduced in x86-64 to force
interrupts to *always* use a new explicit stack. The legacy behaviour is
for a new stack to only be used if switching privilidge levels, meaning
if we got an interrupt in user mode we would get a new stack (prior to
pushing registers), but an interrupt in kernel mode would push registers
directly to the current (kernel) stack. Using the IST allows interrupts
from kernel mode to also switch to a new stack prior to pushing registers,
allowing for much more consistent semantics when dealing with interrupts.