x86: Disable breakpoints in VCPUs
This change allows CONFIG_VTX to work with CONFIG_HARDWARE_BREAKPOINT_SUPPORT
We do not allow for setting hardware breakpoints to be triggered during
guest execution, rather we disable any breakpoints when switching to a VCPU

In addition to disabling any breakpoints before switching to a guest the
hardware will reload DR7 on guest exit to not have any breakpoints set.
Thus the guest is already prevented from setting breakpoints that can get
triggered once we exit the guest. Although our breakpoint state has been
cleared if we then switch to a native thread all of the breakpoint state
will already get reloaded.