arm: SELFOUR-167: Use write allocate cache policy in kernel
Kernel currently uses a write back but not write allocate cache policy for kernel
memory mappings. This results in extremely poor performance for kernel operations
that need to zero memory that is not currently located in the cache. The change
in this commit is to use a write back and write allocate cache policy, matching
the policy used for user mappings