arm: Read/Write fields from active VCPUs
This is a commit to prepare for adding additional registers to the Read
and Write operations. As these registers are expected to be manipulated
in the course of normal virtualization the overhead of effectively switching
the VCPU out, just to modify a register, and then switch it back in, is
excessively high.