ARM/SMP: Implement cpuIndexToID
Currently we assume CPU IDs to be linear and use HW CPU IDs as indices
to SMP-related arrays/state. This may change in the future (when
cluster-based SoC are supported).

cpuIndexToID converts a CPU index to a bit position in the case of ARM/GIC,
in order to send IPIs.